ai-control/requirements.txt
# AI Control
Python dependencies

# Core
python-telegram-bot>=20.7
pyyaml>=6.0

# Automation
pyautogui>=0.9.54
pytesseract>=0.3.10
Pillow>=10.2.0

# AI Integration
requests>=2.31.0

# Web (optional)
fastapi>=0.109.0
uvicorn>=0.27.0
aiofiles>=23.2.1

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.8